+++ lib/src/views/image_fall_view.dart
@@ -11,8 +11,6 @@ import 'package:sakura_blizzard/sakura_blizzard.dart';
 ///
 /// (ja) 指定した画像が落下するビューです。
 ///
-/// Author Masahide Mori
-///
 class ImageFallView extends StatefulWidget {
   final Widget child;
   final Size viewSize;
@@ -27,6 +25,8 @@ class ImageFallView extends StatefulWidget {
   final double minBrightness;
   final int fps;
   final Sp3dPhysics Function()? customPhysicsCreation;
+  final bool showIndicatorOnLoading;
+  final bool enablePositionReset;
 
   /// * [child] : A child view will placed between the front and back layers.
   /// * [viewSize] : This view size.
@@ -52,6 +52,10 @@ class ImageFallView extends StatefulWidget {
   /// and 1 means the brightness will not change.
   /// * [customPhysicsCreation] : You can create your own behavior and use it.
   /// If this is not null, the dropType parameter is ignored.
+  /// * [showIndicatorOnLoading] : If true, an indicator will be displayed
+  /// while loading. If false, the child will be displayed.
+  /// * [enablePositionReset] : Specifies whether the object's position is
+  /// reset after it falls. Set to false for one-time effects.
   const ImageFallView(
       {required this.child,
       required this.viewSize,
@@ -66,6 +70,8 @@ class ImageFallView extends StatefulWidget {
       this.fps = 60,
       this.minBrightness = 0.0,
       this.customPhysicsCreation,
+      this.showIndicatorOnLoading = true,
+      this.enablePositionReset = true,
       super.key});
 
   @override
@@ -138,6 +144,8 @@ class _ImageFallViewState extends State<ImageFallView> {
         minBrightness: widget.minBrightness,
         fps: widget.fps,
         key: UniqueKey(),
+        showIndicatorOnLoading: widget.showIndicatorOnLoading,
+        enablePositionReset: widget.enablePositionReset,
         child: widget.child,
       );
     });